+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fixed a deviation from the paper, where `DeltaFScore` also took into account the agreement in predictions of the negative label. ([#51](https://github.com/webis-de/small-text/pull/51))
+- Fixed a bug in `KappaAverage` that affected the stopping behavior. ([#52](https://github.com/webis-de/small-text/pull/52))
 
 ---